Refactor: Split be/routers/games.py into modular files
